+++ src/libraries/ConstantsLib.sol
@@ -7,6 +7,6 @@ bytes32 constant PERMIT_TYPEHASH =
     keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
 uint256 constant MAX_RATE_PER_SECOND = (1e18 + 200 * 1e16) / uint256(365 days); // 200% APR
 uint256 constant TIMELOCK_CAP = 2 weeks;
-uint256 constant MAX_PERFORMANCE_FEE = 0.5e18; // 50%
-uint256 constant MAX_MANAGEMENT_FEE = 0.05e18 / uint256(365 days); // 5%
+uint96 constant MAX_PERFORMANCE_FEE = 0.5e18; // 50%
+uint96 constant MAX_MANAGEMENT_FEE = 0.05e18 / uint96(365 days); // 5%
 uint256 constant MAX_FORCE_REALLOCATE_TO_IDLE_PENALTY = 0.01e18; // 1%
